Reorganized panel elements to separate folder
