+++ assets/src/edit-story/panels/backgroundColor.js
@@ -11,7 +11,8 @@ import { useEffect, useState } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { Panel, PanelTitle, PanelContent, InputGroup, getCommonValue } from './panel';
+import { Panel, PanelTitle, PanelContent } from './panel';
+import { InputGroup, getCommonValue } from './elements';
 
 function BackgroundColorPanel( { selectedElements, onSetProperties } ) {
 	const backgroundColor = getCommonValue( selectedElements, 'backgroundColor' );
