+++ assets/src/edit-story/panels/elements/actionButton.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import styled from 'styled-components';
+
+const ActionButton = styled.button`
+	color: ${ ( { theme } ) => theme.colors.mg.v1 };
+	font-size: 11px;
+`;
+
+export default ActionButton;
