+++ assets/src/edit-story/panels/elements/getCommonValue.js
@@ -1,0 +1,7 @@
+function getCommonValue( list, property ) {
+	const first = list[ 0 ][ property ];
+	const allMatch = list.every( ( el ) => el[ property ] === first );
+	return allMatch ? first : '';
+}
+
+export default getCommonValue;
