+++ assets/src/edit-story/panels/elements/inputGroup.js
@@ -1,0 +1,60 @@
+/**
+ * External dependencies
+ */
+import styled from 'styled-components';
+import PropTypes from 'prop-types';
+
+/**
+ * Internal dependencies
+ */
+import Label from './label';
+import Group from './group';
+
+const Input = styled.input`
+	color: ${ ( { theme } ) => theme.colors.mg.v1 };
+	border: 1px solid;
+	border-radius: 4px;
+	font-size: 11px;
+	line-height: 16px;
+	width: 100px;
+`;
+
+function InputGroup( { type, label, value, isMultiple, onChange, postfix, disabled } ) {
+	const placeholder = isMultiple ? '( multiple )' : '';
+	const isCheckbox = type === 'checkbox';
+	return (
+		<Group disabled={ disabled }>
+			<Label>
+				{ label }
+			</Label>
+			<Input
+				type={ type || 'number' }
+				disabled={ disabled }
+				onChange={ ( evt ) => onChange( isCheckbox ? evt.target.checked : evt.target.value, evt ) }
+				onBlur={ ( evt ) => evt.target.form.dispatchEvent( new window.Event( 'submit' ) ) }
+				placeholder={ placeholder }
+				value={ isCheckbox ? '' : value }
+				checked={ isCheckbox ? value : null }
+			/>
+			{ postfix }
+		</Group>
+	);
+}
+
+InputGroup.propTypes = {
+	type: PropTypes.string,
+	label: PropTypes.string.isRequired,
+	value: PropTypes.any.isRequired,
+	isMultiple: PropTypes.bool.isRequired,
+	onChange: PropTypes.func.isRequired,
+	postfix: PropTypes.string,
+	disabled: PropTypes.bool,
+};
+
+InputGroup.defaultProps = {
+	type: 'number',
+	postfix: '',
+	disabled: false,
+};
+
+export default InputGroup;
