+++ assets/src/edit-story/panels/elements/label.js
@@ -1,0 +1,13 @@
+/**
+ * External dependencies
+ */
+import styled from 'styled-components';
+
+const Label = styled.span`
+	color: ${ ( { theme } ) => theme.colors.mg.v1 };
+	font-size: 11px;
+	line-height: 16px;
+	width: 80px;
+`;
+
+export default Label;
