+++ assets/src/edit-story/panels/panel/index.js
@@ -1,4 +1,3 @@
 export { default as Panel } from './panel';
 export { default as PanelTitle } from './title';
 export { default as PanelContent } from './content';
-export * from './utils';
